requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
python-telegram-bot>=20.0
Pillow>=10.0
